requirements.txt
transformers[torch]
datasets
kaggle
huggingface-hub
python-dotenv
jupyterlab
scikit-learn